Remove a call to variants() helper function
We don't actually need to call variants() inside our returned function,
because that looks up the variants element in the main tailwind config object,
and according to the README, the way to use this lib is to pass the variants directly.

Besides, the destructured variants was shadowing the 'variants' array passed in by
the user, so it was never used.